Fix deadlock if something tries to send a message at exactly the same time as the connection closes.
Basically, this could deadlock because the reactor holds a lock whenever it is inside an event handler
method.  The reactor tries to acquire the same lock when you interact with it from other locations (like
a different thread).  Since we wait in handle_close for the message processor thread to terminate, if
the message processor thread interacts with the reactor (specifically, if it calls schedule_wakeup)
while the handler thread is waiting, neither operation will ever complete.

This fix releases the reactor's internal lock (which ACE appears to give us access to for exactly this
reason) before we start waiting, then reacquires it when we stop waiting to restore it to the state
that the reactor is expecting it to be in.