AndroidConnector API change: allow gatewayId to be passed to constructor.
The security implementation requires that the client know the ID of the gateway
it's connected to (to look up the right key and to verify that the gateway
sends the correct identification in its SERVER_FINISH message).  This allows
scripts using AndroidConnector to specify this parameter if they are connected
to a gateway that's not using the default configuration (gateway name "MainGateway").