Work around library path issues in Ubuntu 11.10.
Ubuntu 11.10 appears to put libraries in /usr/lib/<arch>; this
isn't a problem for gcc as it looks there by default, but causes
problems with MPC because it looks for the presence of libraries
before it builds things, and doesn't know to look in this location.

This can be hacked around by adding /usr/lib/<arch> to libpaths;
we were already doing this for the Data Store plugin in 11.04, but
it seems that the list of libraries for which this is required has
expanded in 11.10.